CAY-1354 fix test to use right types, sql INTEGER is mapped to java Integer and sql BIGINT is mapped to java Long
git-svn-id: https://svn.apache.org/repos/asf/cayenne/main/branches/STABLE-3.0@931927 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/cayenne-jdk1.5-unpublished/src/test/java/org/apache/cayenne/access/DataContextProcedureQueryTest.java
@@ -357,7 +357,7 @@ public class DataContextProcedureQueryTest extends CayenneCase {
         ColumnDescriptor[] columns = new ColumnDescriptor[3];
 
         // read ID as Long, and everything else as default types
-        columns[0] = new ColumnDescriptor("ARTIST_ID", Types.INTEGER, Long.class
+        columns[0] = new ColumnDescriptor("ARTIST_ID", Types.BIGINT, Long.class
                 .getName());
         columns[1] = new ColumnDescriptor("ARTIST_NAME", Types.CHAR, String.class
                 .getName());
